backport vDevelopp on 21.0_consilium / massaction prod customer price

+++ htdocs/societe/price.php
@@ -36,6 +36,12 @@ require_once DOL_DOCUMENT_ROOT.'/core/lib/product.lib.php';
 require_once DOL_DOCUMENT_ROOT.'/product/class/product.class.php';
 require_once DOL_DOCUMENT_ROOT.'/core/lib/company.lib.php';
 require_once DOL_DOCUMENT_ROOT.'/societe/class/societe.class.php';
+//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+require_once DOL_DOCUMENT_ROOT.'/product/class/product.class.php';
+require_once DOL_DOCUMENT_ROOT.'/core/class/html.formcompany.class.php';
+require_once DOL_DOCUMENT_ROOT.'/product/class/html.formproduct.class.php';
+require_once DOL_DOCUMENT_ROOT.'/core/class/html.formfile.class.php';
+//////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 /**
  * @var Conf $conf
@@ -64,6 +70,12 @@ $cancel 		= GETPOST('cancel', 'alpha');
 $search_label 	= GETPOST('search_label', 'alpha');
 $search_price 	= GETPOST('search_price');
 $search_price_ttc = GETPOST('search_price_ttc');
+//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+$massaction = GETPOST('massaction', 'alpha');
+$toselect = GETPOST('toselect', 'array');
+$confirm = GETPOST('confirm', 'alpha');
+$i = 0;
+//////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 // Security check
 $socid = GETPOSTINT('socid') ? GETPOSTINT('socid') : GETPOSTINT('id');
@@ -78,13 +90,41 @@ $result = restrictedArea($user, 'societe', $socid, '&societe');
 
 // Initialize objects
 $object = new Societe($db);
+//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+$form = new Form($db);
+//////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 $error = 0;
+//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+// List of mass actions available
+$arrayofmassactions = array();
+if (getDolGlobalString('PRODUIT_CUSTOMER_PRICES')) {
+	$arrayofmassactions['preupdateprice'] = img_picto('', 'edit', 'class="pictofixedwidth"').$langs->trans("UpdatePrice");
+}
+//Add for delete prod_cust_price
+//if ($user->hasRight('produit', 'supprimer')) {
+//	$arrayofmassactions['predelete'] = img_picto('', 'delete', 'class="pictofixedwidth"').$langs->trans("Delete");
+//}
+
+$massactionbutton = $form->selectMassAction('', $arrayofmassactions);
+if (in_array($massaction, array('predelete', 'preupdateprice'))) {
+	$arrayofmassactions = array();
+}
+//////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 
 /*
  * Actions
  */
+//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+if (GETPOST('cancel', 'alpha')) {
+	$action = 'list';
+	$massaction = '';
+}
+if (!GETPOST('confirmmassaction', 'alpha') && $massaction != 'presend' && $massaction != 'confirm_presend') {
+	$massaction = '';
+}
+//////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 $parameters = array('id' => $socid);
 $reshook = $hookmanager->executeHooks('doActions', $parameters, $object, $action); // Note that $action and $object may have been modified by some hooks
@@ -187,6 +227,20 @@ if (empty($reshook)) {
 		}
 	}
 
+	//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+	$objectclass = 'Product';
+	$permissiontoread = $user->hasRight('produit', 'lire');
+	$permissiontodelete = $user->hasRight('produit', 'supprimer');
+	$permissiontoadd = $user->hasRight('produit', 'creer');
+	$uploaddir = $conf->product->dir_output;
+	include DOL_DOCUMENT_ROOT.'/core/actions_massactions.inc.php';
+
+
+	if ($action == 'delete' && $confirm == 'yes') {
+		$action == 'delete_customer_price';
+	}
+	//////////////////////////////////////////////////////////////////////////////////////////////////////////
+
 	if ($action == 'delete_customer_price' && ($user->hasRight('produit', 'creer') || $user->hasRight('service', 'creer'))) {
 		// Delete price by customer
 		$prodcustprice->id = GETPOSTINT('lineid');
@@ -255,6 +309,9 @@ $object = new Societe($db);
 
 $result = $object->fetch($socid);
 llxHeader("", $langs->trans("ThirdParty").'-'.$langs->trans('PriceByCustomer'));
+//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+$arrayofselected = is_array($toselect) ? $toselect : array();
+//////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 $head = societe_prepare_head($object);
 
@@ -604,7 +661,10 @@ if (getDolGlobalString('PRODUIT_CUSTOMER_PRICES') || getDolGlobalString('PRODUIT
 
 			print '<table class="noborder centpercent">';
 
-			print '<tr class="liste_titre">';
+			//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+			//print '<tr class="liste_titre">';
+			print '<tr class="liste_titre_filter">';
+			//////////////////////////////////////////////////////////////////////////////////////////////////////////
 			print '<td>'.$langs->trans("Product").'</td>';
 			print '<td>'.$langs->trans('RefCustomer').'</td>';
 			print '<td>'.$langs->trans("AppliedPricesFrom").'</td>';
@@ -643,6 +703,9 @@ if (getDolGlobalString('PRODUIT_CUSTOMER_PRICES') || getDolGlobalString('PRODUIT
 				print $userstatic->getNomUrl(-1);
 				print '</td>';
 				print '<td></td>';
+				//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+				$i++;
+				//////////////////////////////////////////////////////////////////////////////////////////////////////////
 			}
 			print "</table>";
 		} else {
@@ -723,9 +786,17 @@ if (getDolGlobalString('PRODUIT_CUSTOMER_PRICES') || getDolGlobalString('PRODUIT
 		print '<!-- view specific price for each product -->'."\n";
 
 		// @phan-suppress-next-line PhanPluginSuspiciousParamOrder
-		print_barre_liste($langs->trans('PriceForEachProduct'), $page, $_SERVER['PHP_SELF'], $option, $sortfield, $sortorder, '', count($prodcustprice->lines), $nbtotalofrecords, '');
+		//print_barre_liste($langs->trans('PriceForEachProduct'), $page, $_SERVER['PHP_SELF'], $option, $sortfield, $sortorder, '', count($prodcustprice->lines), $nbtotalofrecords, '');
 
 		print '<form action="'.$_SERVER["PHP_SELF"].'?id='.$object->id.'" method="POST">';
+		//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+		print_barre_liste($langs->trans('PriceForEachProduct'), $page, $_SERVER['PHP_SELF'], $option, $sortfield, $sortorder, $massactionbutton, count($prodcustprice->lines), $nbtotalofrecords, '');
+		include DOL_DOCUMENT_ROOT.'/core/tpl/massactions_pre.tpl.php';
+
+
+		$varpage = empty($contextpage) ? $_SERVER["PHP_SELF"] : $contextpage;
+		$selectedfields = (count($arrayofmassactions) ? $form->showCheckAddButtons('checkforselect', 1) : '');
+		//////////////////////////////////////////////////////////////////////////////////////////////////////////
 		print '<input type="hidden" name="token" value="'.newToken().'">';
 		print '<input type="hidden" name="id" value="'.$object->id.'">';
 		if (!empty($sortfield)) {
@@ -753,10 +824,28 @@ if (getDolGlobalString('PRODUIT_CUSTOMER_PRICES') || getDolGlobalString('PRODUIT
 		if ($search_price_ttc) {
 			$param .= '&search_price_ttc='.urlencode($search_price_ttc);
 		}
-
-		print '<tr class="liste_titre">';
+		//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+		//print '<tr class="liste_titre">';
+		print '<tr class="liste_titre_filter">';
+		//////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 		$colspan = 0;
+		//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+		$totalarray = array();
+		$totalarray['nbfield'] = 0;
+
+
+		// Show line of result
+		$j = 0;
+		print '<tr data-rowid="'.$object->id.'" class="liste_titre_filter">';
+
+		// Action column
+		if (getDolGlobalString('MAIN_CHECKBOX_LEFT_COLUMN')) {
+			print_liste_field_titre($selectedfields, $_SERVER["PHP_SELF"], "", '', '', '', $sortfield, $sortorder, 'center maxwidthsearch ');
+			$totalarray['nbfield']++;
+		}
+
+		//////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 		foreach ($prodcustprice->fields as $key => $val) {
 			if (!empty($arrayfields['t.'.$key]['checked'])) {
@@ -772,13 +861,30 @@ if (getDolGlobalString('PRODUIT_CUSTOMER_PRICES') || getDolGlobalString('PRODUIT
 				}
 			}
 		}
-		print '<td></td>';
+//		print '<td></td>';
 		$colspan++;
+
+		//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+		// Action column
+		if (!getDolGlobalString('MAIN_CHECKBOX_LEFT_COLUMN')) {
+			print_liste_field_titre($selectedfields, $_SERVER["PHP_SELF"], "", '', '', '', $sortfield, $sortorder, 'center maxwidthsearch ');
+			$totalarray['nbfield']++;
+		}
+		//////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 		print '</tr>';
 
 		if (count($prodcustprice->lines) > 0 || $search_prod) {
 			print '<tr class="liste_titre">';
+			//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+			// Print the search button
+			if (getDolGlobalString('MAIN_CHECKBOX_LEFT_COLUMN')) {
+				print '<td class="liste_titre maxwidthsearch">';
+				$searchpicto = $form->showFilterAndCheckAddButtons(0);
+				print $searchpicto;
+			}
+			print '<td class="liste_titre"></td>';
+			//////////////////////////////////////////////////////////////////////////////////////////////////////////
 			print '<td class="liste_titre"><input type="text" class="flat width75" name="search_prod" value="'.$search_prod.'"></td>';
 			print '<td class="liste_titre" ><input type="text" class="flat width75" name="search_label" value="'.$search_label.'"></td>';
 			print '<td class="liste_titre"></td>';
@@ -790,16 +896,25 @@ if (getDolGlobalString('PRODUIT_CUSTOMER_PRICES') || getDolGlobalString('PRODUIT
 			print '<td class="liste_titre"></td>';
 			print '<td class="liste_titre"></td>';
 			print '<td class="liste_titre"></td>';
-			print '<td class="liste_titre"></td>';
+			//print '<td class="liste_titre"></td>';
 			if (!empty($extralabels)) {
 				foreach ($extralabels as $key) {
 					print '<td class="right"></td>';
 				}
 			}
 			// Print the search button
-			print '<td class="liste_titre maxwidthsearch">';
-			$searchpicto = $form->showFilterAndCheckAddButtons(0);
-			print $searchpicto;
+//			print '<td class="liste_titre maxwidthsearch">';
+//			$searchpicto = $form->showFilterAndCheckAddButtons(0);
+//			print $searchpicto;
+			//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+			if (!getDolGlobalString('MAIN_CHECKBOX_LEFT_COLUMN')) {
+				//////////////////////////////////////////////////////////////////////////////////////////////////////////
+				print '<td class="liste_titre maxwidthsearch">';
+				$searchpicto = $form->showFilterAndCheckAddButtons(0);
+				print $searchpicto;
+				//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+			}
+			//////////////////////////////////////////////////////////////////////////////////////////////////////////
 			print '</td>';
 			print '</tr>';
 		}
@@ -813,6 +928,34 @@ if (getDolGlobalString('PRODUIT_CUSTOMER_PRICES') || getDolGlobalString('PRODUIT
 				$userstatic->fetch($line->fk_user);
 
 				print '<tr class="oddeven">';
+				//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+				if (getDolGlobalString('MAIN_CHECKBOX_LEFT_COLUMN')) {
+					print '<td class="nowrap center">';
+					if ($massactionbutton || $massaction) {   // If we are in select mode (massactionbutton defined) or if we have already selected and sent an action ($massaction) defined
+						$selected = 0;
+						if (in_array($line->id ,$arrayofselected)) {
+							$selected = 1;
+						}
+						print '<input id="cb'.$line->id.'" class="flat checkforselect" type="checkbox" name="toselect[]" value="'.$line->id.'"'.($selected ? ' checked="checked"' : '').'>';
+					}
+					print '<a class="paddingleftonly paddingrightonly" href="'.$_SERVER["PHP_SELF"].'?action=showlog_customer_price&token='.newToken().'&socid='.$object->id.'&prodid='.$line->fk_product.'">';
+					print img_info();
+					print '</a>';
+					print ' ';
+					print '<a class="editfielda paddingleftonly paddingrightonly" href="'.$_SERVER["PHP_SELF"].'?action=edit_customer_price&token='.newToken().'&socid='.$object->id.'&lineid='.$line->id.'">';
+					print img_edit('default', 0, 'style="vertical-align: middle;"');
+					print '</a>';
+					print ' ';
+					print '<a class="paddingleftonly paddingrightonly" href="'.$_SERVER["PHP_SELF"].'?action=delete_customer_price&token='.newToken().'&socid='.$object->id.'&lineid='.$line->id.'">';
+					print img_delete('default', 'style="vertical-align: middle;"');
+					print '</a>';
+
+					print '</td>';
+					if (!$i) {
+						$totalarray['nbfield']++;
+					}
+				}
+				//////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 				print '<td class="left">'.$staticprod->getNomUrl(1)."</td>";
 				print '<td class="left">'.$staticprod->label."</td>";
@@ -831,8 +974,11 @@ if (getDolGlobalString('PRODUIT_CUSTOMER_PRICES') || getDolGlobalString('PRODUIT
 				print '</td>';
 
 				// Extrafields
-				$extrafields->fetch_name_optionals_label("product_customer_price");
-				$extralabels = $extrafields->attributes["product_customer_price"]['label'];
+//				$extrafields->fetch_name_optionals_label("product_customer_price");
+//				$extralabels = $extrafields->attributes["product_customer_price"]['label'];
+				//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+				$staticprod->label = $line->ref_customer;
+				//////////////////////////////////////////////////////////////////////////////////////////////////////////
 				if (!empty($extralabels)) {
 					$sql  = "SELECT";
 					$sql .= " fk_object";
@@ -862,19 +1008,40 @@ if (getDolGlobalString('PRODUIT_CUSTOMER_PRICES') || getDolGlobalString('PRODUIT
 				}
 				// Action
 				if ($user->hasRight('produit', 'creer') || $user->hasRight('service', 'creer')) {
-					print '<td class="right nowraponall">';
-					print '<a class="paddingleftonly paddingrightonly" href="'.$_SERVER["PHP_SELF"].'?action=showlog_customer_price&token='.newToken().'&socid='.$object->id.'&prodid='.$line->fk_product.'">';
-					print img_info();
-					print '</a>';
-					print ' ';
-					print '<a class="editfielda paddingleftonly paddingrightonly" href="'.$_SERVER["PHP_SELF"].'?action=edit_customer_price&token='.newToken().'&socid='.$object->id.'&lineid='.$line->id.'">';
-					print img_edit('default', 0, 'style="vertical-align: middle;"');
-					print '</a>';
-					print ' ';
-					print '<a class="paddingleftonly paddingrightonly" href="'.$_SERVER["PHP_SELF"].'?action=delete_customer_price&token='.newToken().'&socid='.$object->id.'&lineid='.$line->id.'">';
-					print img_delete('default', 'style="vertical-align: middle;"');
-					print '</a>';
-					print '</td>';
+					//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+					if (!getDolGlobalString('MAIN_CHECKBOX_LEFT_COLUMN')) {
+						//////////////////////////////////////////////////////////////////////////////////////////////////////////
+						print '<td class="right nowraponall">';
+						print '<a class="paddingleftonly paddingrightonly" href="'.$_SERVER["PHP_SELF"].'?action=showlog_customer_price&token='.newToken().'&socid='.$object->id.'&prodid='.$line->fk_product.'">';
+						print img_info();
+						print '</a>';
+						print ' ';
+						print '<a class="editfielda paddingleftonly paddingrightonly" href="'.$_SERVER["PHP_SELF"].'?action=edit_customer_price&token='.newToken().'&socid='.$object->id.'&lineid='.$line->id.'">';
+						print img_edit('default', 0, 'style="vertical-align: middle;"');
+						print '</a>';
+						print ' ';
+						print '<a class="paddingleftonly paddingrightonly" href="'.$_SERVER["PHP_SELF"].'?action=delete_customer_price&token='.newToken().'&socid='.$object->id.'&lineid='.$line->id.'">';
+						print img_delete('default', 'style="vertical-align: middle;"');
+						print '</a>';
+						//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+						print '&nbsp;';
+
+
+						if ($massactionbutton || $massaction) {   // If we are in select mode (massactionbutton defined) or if we have already selected and sent an action ($massaction) defined
+							$selected = 0;
+							if (in_array($line->id, $arrayofselected)) {
+								$selected = 1;
+							}
+							print '<input id="cb' . $line->id . '" class="flat checkforselect" type="checkbox" name="toselect[]" value="' . $line->id . '"' . ($selected ? ' checked="checked"' : '') . '>';
+						}
+						if (!$i) {
+							$totalarray['nbfield']++;
+						}
+						//////////////////////////////////////////////////////////////////////////////////////////////////////////
+						print '</td>';
+						//////////////////////// BACKPORT CONSILIUM * DELETE me on V23 //////////////////////////////////////////
+					}
+					//////////////////////////////////////////////////////////////////////////////////////////////////////////
 				}
 
 				print "</tr>\n";
